refactor: separate logging from sending logic to adhere to SRP
